Changes to be committed: 	modified:   PippyUsage.py 	modified:   example.py 	modified:   pippy/IR.py 	modified:   pippy/PipelineDriver.py 	modified:   timm/models/memfuser.py Key modifications: In the initial function of Class Pipe of IR.py
The pipe cannot be pickled because the split_gm is not picklable.
By iterating through the submodules of the split_gm, we check out the submoud_0 has a unnamed module, which is not supported by pickle and may cause an runtime error. Hence, we add the inspect_submodule function to deal with this case.

The other reason for the error is that the split_gm.forward is set to throw. It is defined in this initial function and after set split_gm.forward to throw, the split_gm is not picklable, and the termial raises an error "AttributeError: Can't pickle local object 'Pipe.__init__.<locals>.throw'". To solve this problem, we move throw function to the top of this file.

Combining two of the above reasons, the pipe can be pickled successfully.